+++ tests/lib/test_math.py
@@ -269,6 +269,30 @@ def test_symmetrize_odd():
     assert np.all(A_sym == 5)
 
 
+def test_symmetrize_parity_even():
+    """Tests symmetrization for even parity."""
+    A_sym = maicos.lib.math.symmetrize(np.arange(11).astype(float), is_odd=False)
+    assert np.all(A_sym == 5)
+
+
+def test_symmetrize_parity_odd():
+    """Tests symmetrization for odd parity."""
+    A = np.arange(10).astype(float)
+    A_result = np.arange(10).astype(float) - 4.5
+    A_sym = maicos.lib.math.symmetrize(A, is_odd=True)
+    assert np.all(A_sym == A_result)
+
+
+def test_symmetrize_parity_odd_antisymmetric():
+    """Tests symmetrization for odd parity.
+
+    The array is unchanged, as it is already antisymmetric.
+    """
+    A = np.arange(11).astype(float) - 5
+    A_sym = maicos.lib.math.symmetrize(A, is_odd=True)
+    assert np.all(A_sym == A)
+
+
 def test_higher_dimensions_length_1():
     """Tests arrays with higher dimensions of length 1."""
     A = np.arange(11).astype(float)[:, np.newaxis]