+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Unreleased
 ----------
 Philip Loche, Henrik Stooß
 
+- Introduce `sym_odd` option to `ProfilePlanarBase` to support vectorial
+  observables (#495)
 - Base Cromer-Mann form factor on ``atomgroup.elements`` (#492)
 - Add electron density weights for density modules (#493)
 - Use gallery view for examples (#491)